requirements.txt
Django==1.7.5
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
django-pipeline==1.3.25
djangorestframework==2.4.3
futures==2.2.0
gunicorn==19.1.1
itsdangerous==0.24
psycopg2==2.5.4
wsgiref==0.1.2
django-sslify>=0.2
django-email-as-username>=1.6.7
django-encrypted-fields==1.1.1
xlwt==0.7.5
django-reversion==1.8.5